+++ indexer/sql/src/lib.rs
@@ -7,10 +7,10 @@ pub mod dataloaders;
 pub mod entity;
 pub mod error;
 pub mod hooks;
+pub mod indexer;
 pub mod indexer_path;
 #[cfg(feature = "metrics")]
 pub mod metrics;
-pub mod non_blocking_indexer;
 pub mod pubsub;
 #[cfg(feature = "async-graphql")]
 pub mod scalars;
@@ -19,5 +19,5 @@ pub mod serde_iso8601;
 pub use {
     context::Context,
     error::{IndexerError, Result},
-    non_blocking_indexer::{IndexerBuilder, NonBlockingIndexer},
+    indexer::{IndexerBuilder, NonBlockingIndexer},
 };
